update ingress template to properly handle root context
When `graphdb.workbench.subpath` is set to `/`, the ingress definition doesn't corectly handle the path.
Specifically, the regex in the ingress path needs to be different when the path is /.
E.g. only glob what's behind the slash and use match group $1 in rewrite targer annotation.

Also, `graphdb.external-url` seems to be required when deploying single instance.

GDB-7691