Added new license configurations
Added new configuration properties for the license that aim to improve the configurability if needed.
The new properties are as follows:

  - Added `license.mountPath` to configure where the license volume is mounted
  - Added `license.optional` to configure the license volume as optional if needed
  - Added `license.readOnly` to configure the read/write mode of the license volume mount

Additionally, changed the license directory to `/opt/graphdb/home/conf/license/` with
`license.mountPath` in order to avoid using a `subPath` volume mount. This allows
kubelet to update the license when the Secret has been updated.

Finally, changed the license volume mount as read-only by default with `license.readOnly`