Add option to create indices from existingConfigmap
Currently the Helm Chart supports creating new repositories via an
existing ConfigMap. For deploying GraphDB as IaC this is quite useful.
Unfortunately, there is no way to create indices at install time.

This adds the option to provide a existing ConfigMap via values and
create an index for one or multiple repositories.

The ConfigMap has to consist out of keys that have the same name as the
repository. The value is the SPARQL query.