feat: add support to render string Helm templates in `extraObjects`
The chart can now render string templates added in `extraObjects` array in values of the chart. This change will allow users to also use helm templating functions inside the extra objects providing them more functionality.

The render loop first test if the item in array is a string or not. If it's not a string then it's processed via `toYaml` first

An example values.yaml is also added to show how this feature will work